+++ docs/sapnwrfc.stubs.php
@@ -10,20 +10,20 @@ die('This file is intended for IDE parsers and should not be included.');
 
 /**
  * Clear the function desc cache for $functionName.
- * 
+ *
  * Clears the default repository. If $repositoryId is passed, the cache
  * repository matching $repository is cleard as well.
- * 
+ *
  * @since 1.3.0
- * 
+ *
  * @param string $functionName Function name.
  * @param string $repositoryId The cache repository to use. Defaults to null.
  *                             If passed, the repository is cleared in addition
  *                             to the default repository.
- * 
+ *
  * @return bool True if the call succeeded (this does not necessarily mean
  *              that a cache entry was found).
- * 
+ *
  */
 function clearFunctionDescCache(string $functionName, string $repositoryId = null): bool {}
 
@@ -48,7 +48,7 @@ class Exception extends \RuntimeException
      *
      * @returns array|null
      */
-    public function getErrorInfo(): array {}
+    public function getErrorInfo(): ?array {}
 }
 
 class ConnectionException extends Exception {}
@@ -94,14 +94,15 @@ class Connection
     /**
      * Lookup a RFC function and return a RemoteFunction object.
      *
-     * @param string $functionnName Name of the function.
+     * @param string $functionnName   Name of the function.
+     * @param bool   $invalidateCache If true, invalidates the function desc cache.
      *
      * @return RemoteFunction A RemoteFunction class for the RFC function.
      *
      * @throws FunctionCallException if the lookup fails or an error is
      *                               returned during parameter parsing.
      */
-    public function getFunction(string $functionName): RemoteFunction {}
+    public function getFunction(string $functionName, bool $invalidateCache = false): RemoteFunction {}
 
     /**
      * Retrieve a SSO ticket from the connection.
@@ -159,7 +160,7 @@ class Connection
      *
      * @param string $path Path to trace directory (must exist).
      *
-     * @return bool True if path was set.
+     * @return true True if path was set (BC).
      *
      * @throws ConnectionException if path could not be set.
      */
@@ -170,7 +171,7 @@ class Connection
      *
      * @param int $level Trace level.
      *
-     * @return bool True if level was set.
+     * @return true True if trace level was set (BC).
      *
      * @throws ConnectionException if level could not be set.
      */
@@ -214,7 +215,7 @@ class RemoteFunction
      *
      * @throws FunctionCallException if the parameter status could not be set.
      */
-    public function setParameterActive(string $parameterName, bool $isActive) {}
+    public function setParameterActive(string $parameterName, bool $isActive): void {}
 
     /**
      * Check if a parameter is active or inactive.
@@ -227,9 +228,9 @@ class RemoteFunction
 
     /**
      * Get the function's parameter description.
-     * 
+     *
      * @return array The parameter descriptions, indexed by parameter name.
-     * 
+     *
      * @throws FunctionCallException if the interface description cannot be retrieved.
      */
     public function getFunctionDescription(): array {}