+++ php_sapnwrfc.h
@@ -13,7 +13,7 @@
 extern zend_module_entry sapnwrfc_module_entry;
 #define phpext_sapnwrfc_ptr &sapnwrfc_module_entry
 
-#define PHP_SAPNWRFC_VERSION "2.0.0"
+#define PHP_SAPNWRFC_VERSION "2.1.0"
 #ifdef PHP_WIN32
 #    define PHP_SAPNWRFC_API __declspec(dllexport)
 #elif defined(__GNUC__) && __GNUC__ >= 4