+++ CHANGELOG.md
@@ -6,12 +6,10 @@ All notable changes to this project will be documented in this file.
 - Remove `ZEND_ACC_FINAL` from `RemoteFunction` to make it mockable (#133)
 
 ### Fixed
+- Fix: segfault when freeing a manually created `RemoteFunction` object (#133)
 - Fix an issue where looking up a function module fails if a previous `RemoteFunction` object has been unset
 - Fix an invalid read in `rfc_set_table_row`
 - Remove the incorrect stub file
-
-## Fixed
-- Fix: segfault when freeing a manually created `RemoteFunction` object (#133)
 
 ## [2.0.0] - 2024-05-14
 ### Added